refactor: modularize CLI
